Update deployment dependencies and include APIM template
The changes in `app-deploy.yml` ensure that the deployment job now explicitly depends on the `${{ upper(parameters.ShortName) }}DeployApimSolas` job, which was previously commented out.

The changes in `azure-pipelines.yml` ensure that the `deploy-solas-apim.yml` template is now included in the pipeline with the necessary parameters, which were previously commented out.