Update deployment stage conditions
Modified the `IATDeploy` stage condition to run if the build succeeded and the source branch is not `refs/heads/develop` and does not start with `refs/heads/dev/`.

Updated the `PreProdDeploy` stage condition to run if the build succeeded and the source branch is either `refs/heads/main` or starts with `refs/heads/release/`.