Add xmlTransformationRules to FileTransform@2 tasks
Updated continuous-deployment.yml, functional-test-process.yml,
and azure-pipelines.yml to include the xmlTransformationRules
parameter with an empty string value ('') in all instances of
the FileTransform@2 task. This ensures XML transformation rules
are explicitly set to an empty value, even when enableXmlTransform
is set to false, targeting appsettings.json files in various
directories and zip files.