Add task to copy OpenAPI spec in build pipeline
A new task has been added to the build-test-publish.yml file to copy the OpenAPI specification for the ENS Monitor. This task uses the `CopyFiles@2` task to copy the `external-notification-service-monitor.openApi.yaml` file from the source directory to the target directory specified. The display name for this task is "Copy OpenAPI spec for deployment (ENS Monitor)".